requirements.txt
# MCP Framework
fastmcp>=2.0.0

# Vector Database
pinecone>=5.0.0

# Google Gemini AI
google-generativeai>=0.8.0

# PDF Processing
pymupdf>=1.23.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
